:art: Move setup checks from lamindb here (#352)
* :art: Move setup checks from lamindb here

* :loud_sound: Better logging for instance close

* :recycle: More improvements

* :building_construction: Move upath back from lndb-storage here

* :green_heart: Fix upath imports

* :green_heart: Fix infer_filesystem

* :memo: Prettify docs

* :heavy_minus_sign: Remove pinning of lndb_storage

+++ lndb/_close.py
@@ -1,5 +1,8 @@
 import os
 
+from lamin_logger import logger
+
+from ._settings import settings
 from .dev._settings_store import current_instance_settings_file
 
 
@@ -8,8 +11,10 @@ def close() -> None:
 
     Returns `None` if succeeds, otherwise an exception is raised.
     """
-    try:
+    if current_instance_settings_file().exists():
+        instance = settings.instance.identifier
         current_instance_settings_file().unlink()
         os.environ["LAMINDB_INSTANCE_LOADED"] = "0"
-    except FileNotFoundError:
-        raise
+        logger.success(f"Closed {instance}")
+    else:
+        logger.info("No instance loaded")
